commit 5 calculadora de notas

+++ public/js/script.js
@@ -1,6 +1,6 @@
 function calculoMedia() {
-    let nota1 = parseInt(document.getElementById('nota1').value);
-    let nota2 = parseInt(document.getElementById('nota2').value);
+    let nota1 = parseFloat(document.getElementById('nota1').value);
+    let nota2 = parseFloat(document.getElementById('nota2').value);
     let mediaNotas = (nota1 + nota2) / 2
     if (mediaNotas >= 7) {
         let media1 = document.getElementById("textomedia");
